feat(view-stops-in-map): move nearby stops map to components directory
